refactor: complete code base
files linting

fix: scroll to bottom part

Fix: reusability of cells

Fix: hiding of imageView
