+++ Podfile
@@ -9,11 +9,11 @@ target 'Susi' do
   pod 'Material'
   pod 'Toast-Swift', '~> 2.0.0'
   pod 'Alamofire', '~> 4.4'
-  pod 'RealmSwift'
   pod 'Popover'
   pod 'SwiftValidators'
   pod 'ALTextInputBar'
   pod 'AlamofireImage', '~> 3.1'
   pod 'SwiftLint'
   pod 'Nuke-Gifu-Plugin'
+  pod 'BouncyLayout'
 end
