+++ Susi/Controllers/AppNavigationController.swift
@@ -17,9 +17,9 @@ class AppNavigationController: NavigationController {
         guard let v = navigationBar as? NavigationBar else {
             return
         }
-
         v.depthPreset = .none
         v.dividerColor = UIColor.defaultColor()
         v.backgroundColor = UIColor.defaultColor()
     }
+
 }
