+++ Susi/Custom Views/Message Cells/ChatMessageCell.swift
@@ -1,43 +1,23 @@
 //
-//  AnswerMessageCell.swift
+//  ChatMessageCell.swift
 //  Susi
 //
 //  Created by Chashmeet Singh on 2017-03-21.
 //  Copyright © 2017 FOSSAsia. All rights reserved.
 //
+
 import UIKit
-import Material
-import MapKit
-import Alamofire
-import AlamofireImage
-import NukeGifuPlugin
 
-class ChatMessageCell: BaseCell, MKMapViewDelegate {
+class ChatMessageCell: BaseCell {
 
-    var message: Message? {
-        didSet {
-            messageTextView.text = message?.body
-            websiteText.text = message?.websearchData?.info
+    static let grayBubbleImage = UIImage(named: "bubble_gray")!.resizableImage(withCapInsets: UIEdgeInsets(top: 22, left: 26, bottom: 22, right: 26)).withRenderingMode(.alwaysTemplate)
 
-            if let imageString = message?.websearchData?.image {
-                if let url = URL(string: imageString) {
-                    let urlRequest = URLRequest(url: url)
-                    searchImageView.af_setImage(withURLRequest: urlRequest)
-                }
-            }
-
-            if let imageString = message?.body, imageString.isImage() {
-                if let url = URL(string: imageString) {
-                    AnimatedImage.manager.loadImage(with: url, into: imageView)
-                }
-            }
-        }
-    }
+    static let blueBubbleImage = UIImage(named: "bubble_blue")!.resizableImage(withCapInsets: UIEdgeInsets(top: 22, left: 26, bottom: 22, right: 26)).withRenderingMode(.alwaysTemplate)
 
     let messageTextView: UITextView = {
         let textView = UITextView()
         textView.font = UIFont.systemFont(ofSize: 18)
-        textView.text = "Sample message"
+        textView.text = ControllerConstants.defaultMessage
         textView.isEditable = false
         textView.backgroundColor = .clear
         textView.isSelectable = false
@@ -50,40 +30,6 @@ class ChatMessageCell: BaseCell, MKMapViewDelegate {
         view.layer.masksToBounds = true
         return view
     }()
-
-    let mapView: MKMapView = {
-        let mapView = MKMapView()
-        mapView.mapType = .standard
-        mapView.isZoomEnabled = true
-        mapView.isScrollEnabled = false
-        return mapView
-    }()
-
-    let websearchContentView = UIView()
-
-    let searchImageView: UIImageView = {
-        let imageView = UIImageView(frame: CGRect(x: 0, y: 0, width: 44, height: 44))
-        imageView.contentMode = .scaleAspectFit
-        imageView.image = UIImage(named: "no-image")
-        return imageView
-    }()
-
-    let websiteText: UILabel = {
-        let label = UILabel()
-        label.textColor = UIColor(red:0.26, green:0.26, blue:0.26, alpha:1.0)
-        label.numberOfLines = 2
-        return label
-    }()
-
-    let imageView: AnimatedImageView = {
-        let imageView = AnimatedImageView()
-        imageView.contentMode = .scaleAspectFill
-        imageView.layer.cornerRadius = 8.0
-        return imageView
-    }()
-
-    static let grayBubbleImage = UIImage(named: "bubble_gray")!.resizableImage(withCapInsets: UIEdgeInsets(top: 22, left: 26, bottom: 22, right: 26)).withRenderingMode(.alwaysTemplate)
-    static let blueBubbleImage = UIImage(named: "bubble_blue")!.resizableImage(withCapInsets: UIEdgeInsets(top: 22, left: 26, bottom: 22, right: 26)).withRenderingMode(.alwaysTemplate)
 
     let bubbleImageView: UIImageView = {
         let imageView = UIImageView()
@@ -102,55 +48,6 @@ class ChatMessageCell: BaseCell, MKMapViewDelegate {
         textBubbleView.addConstraintsWithFormat(format: "H:|[v0]|", views: bubbleImageView)
         textBubbleView.addConstraintsWithFormat(format: "V:|[v0]|", views: bubbleImageView)
 
-    }
-
-    func addMapView(_ frame: CGRect) {
-        websearchContentView.removeFromSuperview()
-        imageView.removeFromSuperview()
-        textBubbleView.addSubview(mapView)
-
-        mapView.frame = frame
-
-        if let mapData = message?.mapData {
-            let latitude = mapData.latitude
-            let longitude = mapData.longitude
-
-            let center = CLLocationCoordinate2D(latitude: latitude, longitude: longitude)
-            let zoomLevel = 360 / pow(2, Double(mapData.zoom)) * Double(frame.width) / 256
-            let span = MKCoordinateSpanMake(0, zoomLevel)
-            let region = MKCoordinateRegion(center: center, span: span)
-
-            mapView.setRegion(region, animated: true)
-        }
-    }
-
-    func addLinkPreview(_ frame: CGRect) {
-        mapView.removeFromSuperview()
-        imageView.removeFromSuperview()
-
-        textBubbleView.addSubview(websearchContentView)
-        websearchContentView.backgroundColor = UIColor(red:0.88, green:0.88, blue:0.88, alpha:1.0)
-        websearchContentView.frame = frame
-
-        websearchContentView.addSubview(searchImageView)
-        websearchContentView.addSubview(websiteText)
-        websearchContentView.addConstraintsWithFormat(format: "H:|-4-[v0(44)]-4-[v1]-4-|", views: searchImageView, websiteText)
-        websearchContentView.addConstraintsWithFormat(format: "V:|-4-[v0]-4-|", views: searchImageView)
-        websearchContentView.addConstraintsWithFormat(format: "V:|-4-[v0(44)]-4-|", views: websiteText)
-    }
-
-    func addImageView() {
-        mapView.removeFromSuperview()
-        websearchContentView.removeFromSuperview()
-        messageTextView.text = ""
-        textBubbleView.addSubview(imageView)
-        textBubbleView.addConstraintsWithFormat(format: "H:|-20-[v0]-15-|", views: imageView)
-        textBubbleView.addConstraintsWithFormat(format: "V:|-15-[v0]-15-|", views: imageView)
-    }
-
-    override func prepareForReuse() {
-        super.prepareForReuse()
-        imageView.prepareForReuse()
     }
 
 }
