+++ Susi/Helpers/Color.swift
@@ -18,4 +18,16 @@ extension UIColor {
         return UIColor(red: 0.13, green: 0.13, blue: 0.13, alpha: 1.0)
     }
 
+    class func incomingMessageTintColor() -> UIColor {
+        return UIColor(red: 0.86, green: 0.97, blue: 0.78, alpha: 1.0)
+    }
+
+    class func webTextColor() -> UIColor {
+        return UIColor(red: 0.26, green: 0.26, blue: 0.26, alpha: 1.0)
+    }
+
+    class func linkPreviewBackgroundColor() -> UIColor {
+        return UIColor(red: 0.88, green: 0.88, blue: 0.88, alpha: 1.0)
+    }
+
 }
