+++ Susi/Model/Message.swift
@@ -59,7 +59,6 @@ struct Message {
                         let responseType = response[Client.ChatKeys.ResponseType]
 
                         if responseType == ResponseTypes.answer.rawValue {
-                            debugPrint("Inside Answer")
 
                             if let expression = response[Client.ChatKeys.Expression] {
                                 if expression.isURL() && expression.isImage() {
@@ -69,7 +68,7 @@ struct Message {
                             }
 
                         } else if responseType == ResponseTypes.map.rawValue {
-                            debugPrint("Inside Map")
+
                             self.responseType = ResponseTypes.map
 
                             if let latitude = response[Client.ChatKeys.Latitude],
@@ -82,12 +81,12 @@ struct Message {
                             self.body = self.body?.components(separatedBy: " ").dropLast().joined(separator: " ")
 
                         } else if responseType == ResponseTypes.websearch.rawValue {
-                            debugPrint("Inside Websearch")
+
                             self.responseType = ResponseTypes.websearch
                             self.query = response[Client.ChatKeys.Query]
 
                         } else {
-                            debugPrint("Inside None")
+                            debugPrint("error")
                         }
                     }
